+++ packages/javascript-kernel/src/comm/index.ts
@@ -1,0 +1,4 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+export * from './manager';
