+++ packages/javascript-kernel/src/widgets/version.ts
@@ -1,0 +1,10 @@
+// Copyright (c) Jupyter Development Team.
+// Distributed under the terms of the Modified BSD License.
+
+export const WIDGET_PROTOCOL_VERSION = '2.1.0';
+export const BASE_MODULE = '@jupyter-widgets/base';
+export const BASE_MODULE_VERSION = '2.0.0';
+export const CONTROLS_MODULE = '@jupyter-widgets/controls';
+export const CONTROLS_MODULE_VERSION = '2.0.0';
+export const OUTPUT_MODULE = '@jupyter-widgets/output';
+export const OUTPUT_MODULE_VERSION = '1.0.0';
